+++ v2/pink-sb/src/lib/index.ts
@@ -23,7 +23,7 @@ export { default as Pagination } from './Pagination.svelte';
 export { default as Code } from './Code.svelte';
 export { default as InlineCode } from './InlineCode.svelte';
 export { default as Popover } from './Popover.svelte';
-export { default as Box } from './Box.svelte';
+export { default as Fieldset } from './Fieldset.svelte';
 export { Toast } from './toast/index.js';
 export * as Button from './button/index.js';
 export * as Link from './link/index.js';
